Add get_lines endpoint at /lines
